requirements.txt
absl-py==0.14.1
aiohttp==3.7.4
alembic==1.4.1
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
cloudpickle==2.0.0
databricks-cli==0.15.0
docker==5.0.3
entrypoints==0.3
Flask==2.0.2
fsspec==2021.10.0
future==0.18.2
fvcore==0.1.5.post20210924
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
iopath==0.1.9
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
mlflow==1.20.2
multidict==5.2.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
portalocker==2.3.2
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
pytorch-lightning==1.4.1
pytz==2021.3
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.25
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
torch==1.9.1
torchmetrics==0.5.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
yacs==0.1.8
yarl==1.7.0
zipp==3.6.0

python-dotenv==0.19.2
